Fix error in “Update Gutter” for non-UTF-8 encodings
When the document uses an encoding different from UTF-8, the output of `diff_cmd` may contain characters that are not valid UTF-8. This leads to following error:

    Update Gutter:38:in `block in <main>': invalid byte sequence in UTF-8 (ArgumentError)
    	from Update Gutter:35:in `open'
    	from Update Gutter:35:in `<main>'

This is especially annoying because this error pops up every time the document is saved (because the “Update Gutter” command runs on every save).

A more “correct” approach would be to determine the document encoding and tell Ruby to use this encoding for `io`. But I’m not sure how to do this, and the code only looks at the diff markers at the beginning of lines anyway.