Fix error in “Update Gutter” for non-UTF-8 encodings
When the document uses an encoding different from UTF-8, the output of `diff_cmd` may contain characters that are not valid UTF-8. This leads to following error:

    Update Gutter:38:in `block in <main>': invalid byte sequence in UTF-8 (ArgumentError)
    	from Update Gutter:35:in `open'
    	from Update Gutter:35:in `<main>'

This is especially annoying because this error pops up every time the document is saved (because the “Update Gutter” command runs on every save).

Setting `ascii-8bit` encoding directly on the io object works around the issue as we only care about the beginning part of the line not the actual contents. Fixes #7.